requirements.txt
woocommerce
python-dotenv